requirements.txt
numpy
paddlepaddle_gpu>=2.2
Pillow==8.4.0
scipy
wandb